fix: address code review feedback
- Guard prune_old_attempts to skip 'unknown' outcome (let dedicated pruner own it)
- Align retention_pending_escalation_hours to u32 (consistent with other retention fields)
- Update review_ledger schema description to include 'unknown'
- Fix MCP tool count to 14 in docs
- Clarify unknown retention is "7 days after escalation"

+++ src/config.rs
@@ -53,7 +53,7 @@ pub struct Config {
     pub retention_snapshots_days: u32,
     pub retention_tasks_archive_days: u32,
     pub retention_unknown_days: u32,
-    pub retention_pending_escalation_hours: u64,
+    pub retention_pending_escalation_hours: u32,
 
     pub default_project_name: String,
 }
